+++ services/web/server/src/simcore_service_webserver/security/_authz_access_model.py
