+++ services/web/server/src/simcore_service_webserver/security/_utils.py
@@ -6,7 +6,7 @@
 """
 import operator
 
-from ._access_roles import ROLES_PERMISSIONS
+from ._authz_access_roles import ROLES_PERMISSIONS
 
 
 def named_permissions() -> list[str]:
