+++ services/web/server/tests/unit/isolated/test_security__authz_access.py
@@ -15,11 +15,14 @@ from pathlib import Path
 import jsondiff
 import pytest
 from simcore_service_webserver.projects.models import ProjectDict
-from simcore_service_webserver.security._access_model import (
+from simcore_service_webserver.security._authz_access_model import (
     RoleBasedAccessModel,
     check_access,
 )
-from simcore_service_webserver.security._access_roles import ROLES_PERMISSIONS, UserRole
+from simcore_service_webserver.security._authz_access_roles import (
+    ROLES_PERMISSIONS,
+    UserRole,
+)
 
 
 @pytest.fixture
